### 2024-01-02 #6177 @editorialbot
[PRE REVIEW]: goFlux: A user-friendly way to calculate GHG fluxes yourself, regardless of   user experience
**Submitting author:** @qepanna (<a href="http://orcid.org/0000-0002-4888-7232">Karelle Rheault</a>)
**Repository:** https://github.com/Qepanna/goFlux
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
