### 2024-01-02 #6177 @kthyng
Hi @Qepanna and thanks for your submission. A couple of initial items:
* We require an OSI-approved license to review submissions.
* I don't see any tests directory in your repo but we require some way of testing: https://joss.readthedocs.io/en/latest/review_criteria.html#tests
* I worry about the state of the package given your "beta" comment in the readme and docs. This seems to indicate the package isn't ready for review by JOSS.
