+++ .github/workflows/check-standard.yaml
@@ -30,6 +30,7 @@ jobs:
 
     steps:
       - uses: actions/checkout@v3
+        with: ${{ Qepanna/goFlux }}
 
       - uses: ./setup-pandoc
 
