+++ .github/workflows/check-standard.yaml
@@ -2,9 +2,9 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main, master, v2-branch]
+    branches: [main, master]
   pull_request:
-    branches: [main, master, v2-branch]
+    branches: [main, master]
 
 name: R-CMD-check
 
@@ -30,21 +30,22 @@ jobs:
 
     steps:
       - uses: actions/checkout@v3
-        with: ${{ goFlux }}
 
-      - uses: ./setup-pandoc
+      - uses: r-lib/actions/setup-pandoc@v2
 
-      - uses: ./setup-r
+      - uses: r-lib/actions/setup-r@v2
         with:
           r-version: ${{ matrix.config.r }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
 
-      - uses: ./setup-r-dependencies
+      - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           extra-packages: any::rcmdcheck
           needs: check
 
-      - uses: ./check-r-package
+      - uses: r-lib/actions/check-r-package@v2
         with:
-          upload-snapshots: true
+          args: 'c("--no-manual", "--as-cran")'
+          error-on: '"error"'
+          check-dir: '"check"'
