+++ pkgdown.yml
@@ -1,0 +1,6 @@
+pandoc: 2.19.2
+pkgdown: 2.0.7
+pkgdown_sha: ~
+articles: {}
+last_built: 2024-01-03T10:22Z
+
