### 2024-01-03 #6177 @Qepanna
Hi Kristen,

Thank you for considering my software. I have addressed the following:

  *   I added a LICENSE file to the repository
  *   I added the following test files in the .github/workflow directory:
     *   check-standard.yaml
     *   test-coverage.yaml
     *   pkgdown.yaml
     *   pr-commands.yaml
  *   The package is ready and fully functioning, but since it is meant to work with many versions of different instruments, errors that I cannot anticipate may still occur. I have modified the statement from “BETA VERSION” to “FIRST RELEASE”.

Best regards,
Karelle Rheault

Karelle Rheault
PhD candidate

University of Copenhagen
Department of Geosciences and Natural Resource Management
Section of Forest, Nature and Biomass
Rolighedsvej 23
Frederiksberg C

DIR +45 35 32 43 04
***@***.******@***.***>

[Logo for Københavns Universitet]

How we protect personal data<https://informationssikkerhed.ku.dk/english/protection-of-information-privacy/privacy-policy/>
LinkedIn Profile<https://www.linkedin.com/in/karelle-rheault-063842105/?challengeId=AQFSLWms_eH8SgAAAX_gFPGWCFmZ6_X8zLLTJoLuKaedokcA1gR7aPiAHk6XsXDSWeg2UxWeFqKwogNJfkPqmWUQ04uIRcrkrQ&submissionId=047dbc84-f778-e116-d099-3ab935bd880a>

From: Kristen Thyng ***@***.***>
Sent: Tuesday, January 2, 2024 7:42 PM
Cc: Karelle Rheault ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: goFlux: A user-friendly way to calculate GHG fluxes yourself, regardless of user experience (Issue #6177)

Hi @Qepanna<https://github.com/Qepanna> and thanks for your submission. A couple of initial items:

  *   We require an OSI-approved license to review submissions.
  *   I don't see any tests directory in your repo but we require some way of testing: https://joss.readthedocs.io/en/latest/review_criteria.html#tests<https://joss.readthedocs.io/en/latest/review_criteria.html#tests>
  *   I worry about the state of the package given your "beta" comment in the readme and docs. This seems to indicate the package isn't ready for review by JOSS.
