### 2024-01-03 #6177 @kthyng
@Qepanna Regarding tests, I mean I was looking for a tests directory that runs a variety of tests to make sure that the code is performing correctly and accurately. I see you added some workflows, but are there any tests of this nature in the package? All or at least most functionality of the code base should be checked by tests that can be run by you or a user.

Also I'd recommend you remove or heavily reduce the content in your readme rather than say it is out of date. It's typical to put the basics there (brief description of functionality and installation instructions or similar) then point to the full docs that are up to date. Then you don't have any redundancy or getting out of date between two pages.
