+++ README.md
@@ -3,9 +3,6 @@
 
 # goFlux: A user-friendly way to calculate GHG fluxes yourself, regardless of user experience<img src="man/figures/goFlux.png" align="right" width="200"/>
 
-This document is out of date. Please consult [this
-webpage](https://qepanna.quarto.pub/goflux/) instead.
-
 #### FIRST RELEASE
 
 The package is ready to use and fully functional, but errors may still
@@ -98,6 +95,9 @@ calculate fluxes (LM or HM, that is the question. -Shakespeare, 1603),
 quality check the results objectively (hence the no experience needed)
 and obtain high quality flux estimates from static chamber measurements
 (wonderful!).
+
+Look up [this webpage](https://qepanna.quarto.pub/goflux/) for a
+demonstration of the package usage.
 
 ### Import and measurement identification
 
@@ -458,19 +458,7 @@ Finally, after finding the best flux estimates, one can plot the results
 and visually inspect the measurements using the function `flux.plot()`
 and save the plots as pdf using `flux2pdf()`.
 
-## How to use
-
-Here is a simple example on how to use the package with a single raw
-file from LGR gas analyzers.
-
-> FIRST RELEASE
->
-> The package is ready to use and fully functional, but errors may still
-> occur. Please report any issues to the maintainer, Karelle Rheault
-> (<karh@ign.ku.dk>), including script and raw data if necessary. Thank
-> you for helping me in the development of this tool! 🙏
-
-### Installation
+## Installation
 
 To install a package from GitHub, one must first install the package
 `remotes` from the CRAN:
@@ -500,218 +488,9 @@ The functioning of the package depends on many other packages
 `stats`, `stringr`, `tibble`, `tidyr`, `utils`), which will be installed
 when installing `goFlux`.
 
-Troubleshoot problems with `install_github()`:
+### Troubleshoot problems with `install_github()`
 
-- [Warning: package is in use and will not be
-  installed](#warning-package-is-in-use-and-will-not-be-installed)
-- [Error: API rate limit exceeded](#error-api-rate-limit-exceeded)
-
-### Import raw data into R
-
-``` r
-library(goFlux)
-
-# Get the example raw file from the package
-file.path <- system.file("extdata", "LGR/LGR.txt", package = "goFlux")
-
-# Import raw data from an LGR gas analyzer
-?LGR_import
-LGR_imp <- LGR_import(inputfile = file.path)
-```
-
-Note that raw data can also be imported from a multitude of other
-instruments, and example data files are provided for all of them:
-
-- LI-COR: LI-6400, LI-7810, LI-7820, LI-8100, LI-8200 (smart chamber)
-- Los Gatos Research instruments: GLA131-MGGA, GLA132-UGGA and
-  GLA151-N2OM1
-- GAIA2TECH (DMR) automated chamber ECOFlux
-- Picarro: G2508 and G4301
-- Gasmet: DX4015
-- PP-Systems: EGM-5
-
-To import multiple files from a folder, use the wrapper function
-`import2RData()`.
-
-``` r
-# Get help for import functions from the goFlux package
-?DX4015_import
-?EGM5_import
-?G2508_import
-?G4301_import
-?GAIA_import
-?LI6400_import
-?LI7810_import
-?LI7820_import
-?LI8100_import
-?LI8200_import
-?N2OM1_import
-?import2RData
-```
-
-### Manual identification of measurements
-
-In this example, the `start.time` for each measurement (`UniqueID`) was
-noted manually in the field, and are provided in an auxiliary file
-(`LGR_aux.txt`).
-
-``` r
-# Other usefull packages
-require(dplyr)
-require(purrr)
-
[…]
