### 2024-01-08 #6177 @Qepanna
Hi Kristen,

I’m sorry I’m new to GitHub and creating a package. I do not know what you are referring to. One of the workflows performs a “R-CMD-check”, which tests all the examples provided in the functions’ documentations. If this is not what was asked, would you kindly provide me with an example of how to create a test directory and what it should contain.

Thank you for your patience.

In the meantime, I modified the README file to fit your recommendations.

Best regards,
Karelle

From: Kristen Thyng ***@***.***>
Sent: Wednesday, January 3, 2024 4:26 PM
Cc: Karelle Rheault ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: goFlux: A user-friendly way to calculate GHG fluxes yourself, regardless of user experience (Issue #6177)

@Qepanna<https://github.com/Qepanna> Regarding tests, I mean I was looking for a tests directory that runs a variety of tests to make sure that the code is performing correctly and accurately. I see you added some workflows, but are there any tests of this nature in the package? All or at least most functionality of the code base should be checked by tests that can be run by you or a user.

Also I'd recommend you remove or heavily reduce the content in your readme rather than say it is out of date. It's typical to put the basics there (brief description of functionality and installation instructions or similar) then point to the full docs that are up to date. Then you don't have any redundancy or getting out of date between two pages.
