+++ man/EGM5_import.Rd
@@ -45,7 +45,7 @@ The default is \code{prec = c(3, 1, 500)}.}
 the EGM-5. There are 5 different modes available: (1)
 Measure mode; (2) SRC or Custom mode; (3) CPY mode;
 (4) Injection mode; or (5) Static mode. If no value
-specified, the parameters will be named "param1",
+is specified, the parameters will be named "param1",
 "param2", "param3", "param4" and "param5".}
 }
 \value{
