### 2024-01-10 4a7a298
abs(kappa)
