### 2024-01-10 3c6fdab
Deploying to gh-pages from @ Qepanna/goFlux@1919000581ad9361a33b423096a80848669ff739 🚀
