### 2024-01-16 abba33b
add acknowledgements and logo
