### 2024-01-18 3d36d52
add warning for small sample size
