### 2024-01-18 efe69da
add error when identify plot is closed manually
