### 2024-01-18 9fff03f
Return results even in case of warning
