### 2024-01-29 cdb1669
handle multiple errors and warnings with LM.flux and HM.flux
