+++ R/goFlux_nokmax.R
@@ -51,7 +51,7 @@
 #' @param Tcham numerical value; temperature inside the chamber (Celsius).
 #'              Alternatively, provide the column \code{Tcham} in \code{dataframe}
 #'              if \code{Tcham} is different between samples. If \code{Tcham} is
-#'              not provided, normal air temperature (15°C) is used.
+#'              not provided, 15°C is used as default.
 #' @param warn.length numerical value; limit under which a measurement is
 #'                    flagged for being too short (\code{nb.obs < warn.length}).
 #'
@@ -80,27 +80,39 @@
 #' estimate cannot exceed this maximum curvature.
 #'
 #' All flux estimates, including the \code{\link[goFlux]{MDF}}, are
-#' multiplied by a \code{\link[goFlux]{flux.term}} which is used to
-#' correct for water vapor inside the chamber, as well as convert the units to
-#' obtain a term in nmol or
+#' obtained from the multiplication of the slope and the
+#' \code{\link[goFlux]{flux.term}}, which is used to correct the
+#' pressure inside the chamber (due to water vapor), as well as convert the
+#' units to obtain a term in nmol or
 #' \ifelse{html}{\out{µmol m<sup>-2</sup>s<sup>-1</sup>}}{\eqn{µmol m^{-2}s^{-1}}{ASCII}}.
 #'
 #' The argument \code{Area} is in \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}},
 #' but the output units from \code{\link[goFlux]{goFlux}} are in
-#' \ifelse{html}{\out{(m<sup>2</sup>)}}{\eqn{(m^2)}{ASCII}}. This means that there is a factor
-#' of 10,000 to convert from \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}}
+#' \ifelse{html}{\out{(m<sup>2</sup>)}}{\eqn{(m^2)}{ASCII}}. This is due to the
+#' conversion from \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}}
+#' to \ifelse{html}{\out{(m<sup>2</sup>)}}{\eqn{(m^2)}{ASCII}} within the
+#' function. This means that there is a factor of 10,000 to convert from
+#' \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}}
 #' to \ifelse{html}{\out{(m<sup>2</sup>)}}{\eqn{(m^2)}{ASCII}}. This is important
-#' to take into account if one would provide something else than an \code{Area} in
-#' \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}} to the function.
+#' to take into account if one would provide something else than an \code{Area}
+#' in \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}} to the function.
 #' For example, with incubated soil samples, one may provide an amount of soil
-#' (kg) instead of an \code{Area}. To get the right units in that case, multiply the
-#' kilograms of soil by 10,000 to remove the conversion from
-#' \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}} to
+#' (kg) instead of an area in the column \code{Area}. To get the right units in
+#' that case, multiply the kilograms of soil by 10,000 to remove the conversion
+#' from \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}} to
 #' \ifelse{html}{\out{(m<sup>2</sup>)}}{\eqn{(m^2)}{ASCII}}.
 #'
[…]
