+++ R/uCH4_import.R
@@ -1,0 +1,226 @@
+#' Import function for the MIRA Ultra
+#' \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}/
+#' \ifelse{html}{\out{C<sub>2</sub>H<sub>6</sub>}}{\eqn{C[2]H[6]}{ASCII}} analyzer
+#' from Aeris Technologies
+#'
+#' Imports single raw gas measurement files from the MIRA Ultra
+#' \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}/
+#' \ifelse{html}{\out{C<sub>2</sub>H<sub>6</sub>}}{\eqn{C[2]H[6]}{ASCII}}
+#' Natural Gas Leak Detection System (with GPS) from Aeris Technologies
+#' (\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}},
+#' \ifelse{html}{\out{C<sub>2</sub>H<sub>6</sub>}}{\eqn{C[2]H[6]}{ASCII}} and
+#' \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}) with the extension .txt
+#'
+#' @param inputfile character string; the name of a file with the extension .txt
+#' @param date.format character string; specifies the date format found in the
+#'                    raw data file. Choose one of the following: "dmy", "ymd",
+#'                    or "mdy". Default is "mdy", as it is the date format from
+#'                    the example data file provided.
+#' @param timezone character string; a time zone in which to import the data to
+#'                 POSIXct format. Default is "UTC". Note about time zone: it is
+#'                 recommended to use the time zone "UTC" to avoid any issue
+#'                 related to summer time and winter time changes.
+#' @param save logical; if \code{save = TRUE}, saves the file as an .RData file
+#'             in a RData folder in the current working directory. If
+#'             \code{save = FALSE}, returns the file in the Console, or load in
+#'             the Environment if assigned to an object.
+#' @param keep_all logical; if \code{keep_all = TRUE}, keep all columns from the raw
+#'                 file. The default is \code{keep_all = FALSE}, and columns that
+#'                 are not necessary for gas flux calculation are removed.
+#' @param prec numerical vector; the precision of the instrument for each gas,
+#'             in the following order: "CH4_ppb", "C2H6dry_ppb" and "H2O_ppm".
+#'             The default is \code{prec = c(1, 0.5, 1)}. The precision for
+#'             water vapor is currently unknown, and a default of 1 ppm is used.
+#'
+#' @returns A data frame containing raw data from the MIRA Ultra
+#' \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}/
+#' \ifelse{html}{\out{C<sub>2</sub>H<sub>6</sub>}}{\eqn{C[2]H[6]}{ASCII}} analyzer
+#' from Aeris Technologies
+#'
+#' @include goFlux-package.R
+#'
+#' @details
+#' In \code{date.format}, the date format refers to a date found in the raw data
+#' file, not the date format in the file name. For these instruments, the date
+#' is found in the column "Time Stamp".
+#'
+#' Note that this function was designed for the following units in the raw file:
+#' \itemize{
+#'   \item ppm for \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
+#'   \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}
[…]
