+++ man/DX4015_import.Rd
@@ -37,8 +37,8 @@ file. The default is \code{keep_all = FALSE}, and columns that
 are not necessary for gas flux calculation are removed.}
 
 \item{prec}{numerical vector; the precision of the instrument for each gas,
-in the following order: "CO2dry_ppm", "COdry_ppb", CH4dry_ppb",
-"N2Odry_ppb", "NH3dry_ppb" and H2O_ppm". The default is
+in the following order: "CO2dry_ppm", "COdry_ppb", "CH4dry_ppb",
+"N2Odry_ppb", "NH3dry_ppb" and "H2O_ppm". The default is
 \code{prec = c(1.6, 23, 13, 2, 23, 33)}.}
 }
 \value{
@@ -108,7 +108,9 @@ See also, import functions for other instruments:
          \code{\link[goFlux]{LI7820_import}},
          \code{\link[goFlux]{LI8100_import}},
          \code{\link[goFlux]{LI8200_import}},
-         \code{\link[goFlux]{N2OM1_import}}
+         \code{\link[goFlux]{N2OM1_import}},
+         \code{\link[goFlux]{uCH4_import}},
+         \code{\link[goFlux]{uN2O_import}}
 
 See \code{\link[base]{timezones}} for a description of the underlying
          timezone attribute.

[…]
