### 2024-01-30 4438b65
handle errors in loops: skip preblematic files and return all error messages after loop
