+++ R/DX4015_import.R
@@ -26,8 +26,8 @@
 #'                 file. The default is \code{keep_all = FALSE}, and columns that
 #'                 are not necessary for gas flux calculation are removed.
 #' @param prec numerical vector; the precision of the instrument for each gas,
-#'             in the following order: "CO2dry_ppm", "COdry_ppb", CH4dry_ppb",
-#'             "N2Odry_ppb", "NH3dry_ppb" and H2O_ppm". The default is
+#'             in the following order: "CO2dry_ppm", "COdry_ppb", "CH4dry_ppb",
+#'             "N2Odry_ppb", "NH3dry_ppb" and "H2O_ppm". The default is
 #'             \code{prec = c(1.6, 23, 13, 2, 23, 33)}.
 #'
 #' @returns A data frame containing raw data from the Gasmet DX4015 gas analyzer.
@@ -80,7 +80,9 @@
 #'          \code{\link[goFlux]{LI7820_import}},
 #'          \code{\link[goFlux]{LI8100_import}},
 #'          \code{\link[goFlux]{LI8200_import}},
-#'          \code{\link[goFlux]{N2OM1_import}}
+#'          \code{\link[goFlux]{N2OM1_import}},
+#'          \code{\link[goFlux]{uCH4_import}},
+#'          \code{\link[goFlux]{uN2O_import}}
 #'
 #' @seealso See \code{\link[base]{timezones}} for a description of the underlying
 #'          timezone attribute.
@@ -117,153 +119,175 @@ DX4015_import <- function(inputfile, date.format = "ymd", timezone = "UTC",
     COdry_ppm <- NH3dry_ppm <- N2Odry_ppm <- CH4dry_ppm <- P_unit <- . <-
     H2O_unit <- CO2_unit <- CH4_unit <- N2O_unit <- CO_unit <- NH3_unit <-
     COwet_ppm <- NH3wet_ppm <- N2Owet_ppm <- CH4wet_ppm <- CO2wet_ppm <-
-    DATE <- TIME <- NULL
+    DATE <- TIME <- import.error <- NULL
 
-  # Load data file
-  data.raw <- read.delim(inputfile, colClasses = "character") %>%
-    # Modify column names for DATE and TIME
-    rename(DATE = Date, TIME = Time) %>%
-    # Modify column names for units
-    rename(H2O_unit = which(names(.) == "Water.vapor.H2O") +1) %>%
-    rename(CO2_unit = which(names(.) == "Carbon.dioxide.CO2") +1) %>%
-    rename(CH4_unit = which(names(.) == "Methane.CH4") +1) %>%
-    rename(N2O_unit = which(names(.) == "Nitrous.oxide.N2O") +1) %>%
-    rename(CO_unit = which(names(.) == "Carbon.monoxide.CO") +1) %>%
-    rename(NH3_unit = which(names(.) == "Ammonia.NH3") +1) %>%
-    rename(P_unit = which(names(.) == "Ambient.pressure") +1) %>%
-    rename(Int_temp_unit = which(names(.) == "Interferometer.temperature") +1) %>%
-    rename(Detect_temp_unit = which(names(.) == "Detector.temperature") +1) %>%
-    rename(IFG_peak_unit = which(names(.) == "IFG.peak.height") +1) %>%
-    # In Int_temp_unit and IFG_peak_unit, replace non ACSII characters
-    mutate_all(str_replace_all, "\uFFFDC", "Celsius") %>%
-    # Modify column names for gas compensation (wet or dry fraction)
-    rename(H2O_frac = which(names(.) == "Water.vapor.H2O") +2) %>%
-    rename(CO2_frac = which(names(.) == "Carbon.dioxide.CO2") +2) %>%
[…]
