+++ R/LI6400_import.R
@@ -21,8 +21,8 @@
 #'                 file. The default is \code{keep_all = FALSE}, and columns that
 #'                 are not necessary for gas flux calculation are removed.
 #' @param prec numerical vector; the precision of the instrument for each gas,
-#'             in the following order: "CO2dry_ppm" and H2O_ppm". The default is
-#'             \code{prec = c(0.15, 20)}.
+#'             in the following order: "CO2dry_ppm" and "H2O_ppm". The default
+#'             is \code{prec = c(0.15, 20)}.
 #'
 #' @returns A data frame containing raw data from LI-COR GHG analyzer LI-6400.
 #'
@@ -69,7 +69,9 @@
 #'          \code{\link[goFlux]{LI7820_import}},
 #'          \code{\link[goFlux]{LI8100_import}},
 #'          \code{\link[goFlux]{LI8200_import}},
-#'          \code{\link[goFlux]{N2OM1_import}}
+#'          \code{\link[goFlux]{N2OM1_import}},
+#'          \code{\link[goFlux]{uCH4_import}},
+#'          \code{\link[goFlux]{uN2O_import}}
 #'
 #' @seealso See \code{\link[base]{timezones}} for a description of the underlying
 #'          timezone attribute.
@@ -101,115 +103,134 @@ LI6400_import <- function(inputfile, date.format = "mdy", timezone = "UTC",
 
   # Assign NULL to variables without binding
   cham.close <- cham.open <- POSIX.time <- chamID <- DATE <- TIME <- H2O_mmol <-
-    Etime <- CO2dry_ppm <- Meas.type <- plotID <- H2OS <- Cdry <- Press <-
-    Tair <- Mode <- ETime <- HHMMSS <- Meas.type..NEE.ER. <- Plot. <- Obs <-
+    Etime <- CO2dry_ppm <- Meas.type <- plotID <- H2OS <- Cdry <- Obs <- PAR <-
+    Tair <- Mode <- ETime <- HHMMSS <- Meas.type..NEE.ER. <- Plot. <- Press <-
     V4 <- V1 <- start.time <- met.date.warning <- flag <- offset <- Vcham <-
-    Area <- Pcham <- Tcham <- H2O_ppm <- TCham <- PAR <- PARo <- NULL
+    Area <- Pcham <- Tcham <- H2O_ppm <- TCham <- PARo <- import.error <- NULL
 
-  # Find how many rows need to be skipped
-  skip.rows <- tryapply(seq(1:30), function(i) {
-    if (read.delim(inputfile, header = F, skip = i)[1,1] == "$STARTOFDATA$") {
-      return(i) }
-  }) %>% unlist %>% as.numeric() +1
+  # Input file name
+  inputfile.name <- gsub(".*/", "", inputfile)
 
-  # Import metadata from LI6400 (.txt)
-  metadata <- read.delim(inputfile, header = F, comment.char = "<") %>%
-    filter(V1 == "Const=" | V4 == "")
-
-  # Extract date from metadata
-  tryCatch(
-    {if (date.format == "dmy") {
-      try.met.date <- dmy(substr(metadata[2,1], 5, nchar(metadata[2,1])-9), tz = timezone)
-    } else if (date.format == "mdy") {
-      try.met.date <- mdy(substr(metadata[2,1], 5, nchar(metadata[2,1])-9), tz = timezone)
-    } else if (date.format == "ymd") {
-      try.met.date <- ymd(substr(metadata[2,1], 5, nchar(metadata[2,1])-9), tz = timezone)
-    }}, warning = function(w) {met.date.warning <<- "date.format.error"}
+  # Try to load data file
+  try.import <- tryCatch(
[…]

[…]
