### 2024-01-31 473ad17
use devtools instead of remotes to install package from GitHub
