### 2024-02-08 588a93d
When window is closed, identification is skiped. Return warning instead of stoping the loop.
