### 2024-02-08 e23ba7c
fix: save pdf to working directory by default
