### 2024-02-09 e6a4676
add details about new instruments uCH4 and uN2O
