+++ R/import2RData.R
@@ -61,6 +61,8 @@
 #' \item LI-8100: column Date
 #' \item LI-8200: (see comment below)
 #' \item N2OM1: column Time
+#' \item uCH4: column Time Stamp
+#' \item uN2O: column Time Stamp
 #' }
 #' For the instrument LI-6400, the date is found in one of the first lines in
 #' a format containing abbreviations, for example "Thr Aug 6 2020", which would
@@ -86,6 +88,8 @@
 #' @include LI8100_import.R
 #' @include LI8200_import.R
 #' @include N2OM1_import.R
+#' @include uCH4_import.R
+#' @include uN2O_import.R
 #'
 #' @seealso Use the wrapper function \code{\link[goFlux]{import2RData}}
 #'          to import multiple files from the same folder path using any instrument.
@@ -101,7 +105,9 @@
 #'          \code{\link[goFlux]{LI7820_import}},
 #'          \code{\link[goFlux]{LI8100_import}},
 #'          \code{\link[goFlux]{LI8200_import}},
-#'          \code{\link[goFlux]{N2OM1_import}}
+#'          \code{\link[goFlux]{N2OM1_import}},
+#'          \code{\link[goFlux]{uCH4_import}},
+#'          \code{\link[goFlux]{uN2O_import}}
 #'
 #' @examples
 #' # Examples on how to use it with all instruments.
