+++ DESCRIPTION
@@ -45,9 +45,9 @@ Collate:
     'k.max.R'
     'goFlux.R'
     'goFlux_nokmax.R'
+    'uN2O_import.R'
+    'uCH4_import.R'
     'import2RData.R'
     'obs.win.R'
     'seq.rep.R'
-    'uCH4_import.R'
-    'uN2O_import.R'
 Imports: AICcmodavg, data.table, dplyr, ggnewscale, ggplot2, ggstar, graphics, grDevices, grid, gridExtra, lubridate, minpack.lm, msm, pbapply, plyr, purrr, rjson, rlist, SimDesign, stats, stringr, tibble, tidyr, utils
