### 2024-02-09 7e66102
add k.min parameter in goFlux and HM.flux
