### 2024-02-09 62d24eb
add k.min parameter in goFlux and HM.flux
