### 2024-02-09 #6177 @kthyng
Hi @RMeli and welcome to JOSS! Any chance you'd be comfortable enough to edit this submission? I see that it is not immediately in your domain but you are also the only editor with chemistry expertise and this submission has been waiting for awhile. If not, no worries!
