### 2024-02-18 #6177 @Qepanna
Hi @arfon, I believe anyone with the area of expertise "Earth Sciences and Ecology" and knows the language R could be a good candidate. Here is my selection of a few candidates who also have some relevant key words in their topic area:
[@chenyangkang](https://github.com/chenyangkang)
[@shubhamjain15](https://github.com/shubhamjain15)
[@david-yannick](https://github.com/david-yannick)
[@nmstreethran](https://github.com/nmstreethran)
[@hahsan1](https://github.com/hahsan1)

I hope this helps.
