### 2024-02-19 #6177 @arfon
@ibarraespinosa @Schuch666 @dkauf42 @hahsan1 :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _goFlux: A user-friendly way to calculate GHG fluxes yourself, regardless of user experience_

Many thanks
Arfon
