### 2024-02-20 #6177 @ibarraespinosa
hi @arfon , I'm reviewing gcamreport https://github.com/openjournals/joss-reviews/issues/5975#issuecomment-1953097401. Once I have finished I can help with this
