### 2024-02-20 #6177 @dkauf42
Hi @arfon, lacking familiarity with R, I don't think I'm up for reviewing this. Please do keep me (or via my other GitHub handle, @danielfromearth) in mind for other reviews though!
