### 2024-02-21 #6177 @arfon
> Yes it work for me, I'm happy to help with this review.

Thanks @Schuch666 – we ideally would want the reviews to be complete within ~6 weeks which means providing initial feedback in 2-3 weeks. Does that sound OK?

> hi @arfon , I'm reviewing gcamreport https://github.com/openjournals/joss-reviews/issues/5975#issuecomment-1953097401. Once I have finished I can help with this

Thanks @ibarraespinosa – I failed to notice you're working on another review for us right now. I'll keep looking for a second reviewer but may come back to you later if I don't identify anyone else. Thanks!
