### 2024-02-21 #6177 @arfon
@shubhamjain15 @david-yannick @nmstreethran - 👋 would any of you be willing to review this submission for JOSS? The submission under consideration is goFlux: A user-friendly way to calculate GHG fluxes yourself, regardless of user experience

Many thanks
Arfon
