### 2024-03-01 9388b2a
EGM5 now accepts two file formats: .TXT and .csv
