+++ R/import2RData.R
@@ -276,8 +276,8 @@ import2RData <- function(path, instrument, date.format, timezone = "UTC",
   if(instrument == "EGM5"){
 
     # List all the files contained in the specified path
-    file_list <- list.files(path = path, pattern = "\\.TXT", recursive = T,
-                            full.names = TRUE)
+    file_list <- list.files(path = path, pattern = "\\.TXT|\\.csv",
+                            recursive = T, full.names = TRUE)
 
     # Loop through files in "file_list" and apply import functions
     pblapply(seq_along(file_list), function(i) {
