+++ pkgdown.yml
@@ -1,6 +1,6 @@
-pandoc: 2.19.2
+pandoc: 3.1.11
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-02-09T08:18Z
+last_built: 2024-03-01T07:58Z
 
