### 2024-03-01 #6393 @Schuch666
Hi @Qepanna, You mention in the installation section that "must first install the package devtools", but the user can use another package that provide the same functionality such as the `remotes` package:

`remotes::install_github('qepanna/goflux')`

can you include a note that include remotes?
