### 2024-03-02 #6393 @shubhamjain15
Hi @Qepanna, Could you take a look at this error in the 'uCH4_import' and 'uN2O_import' functions. My guess is you need to remove the highlighted lines (attached image) in the code which may have been for testing purposes only.

```{r}
library(goFlux)
file.path <- system.file("extdata", "uCH4/uCH4.txt", package = "goFlux")
uCH4_imp <- uCH4_import(inputfile = file.path)
file.path <- system.file("extdata", "uN2O/uN2O.txt", package = "goFlux")
uN2O_imp <- uN2O_import(inputfile = file.path)
```
Warning: cannot open file 'inst/extdata/uCH4/uCH4.txt': No such file or directoryWarning: Error occurred in file uCH4.txt:
   Error in file(file, "rt"): cannot open the connection
Warning: cannot open file 'inst/extdata/uN2O/uN2O.txt': No such file or directoryWarning: Error occurred in file uN2O.txt:
   Error in file(file, "rt"): cannot open the connection

![import_error](https://github.com/openjournals/joss-reviews/assets/40867903/d85a5b88-12da-4ec1-8519-af530ac2ad27)
