### 2024-03-05 #6393 @Qepanna
> Hi @Qepanna, You mention in the installation section that "must first install the package devtools", but the user can use another package that provide the same functionality such as the `remotes` package:
>
> `remotes::install_github('qepanna/goflux')`
>
> can you include a note that include remotes?

Hi @Schuch666, originally it was included that one could use the remotes packages, however I encounterd some errors using this package and I thought it would be best to avoid using it. Would you still recommend that I include the package remotes?
