### 2024-03-05 #6393 @Qepanna
> @Qepanna In the 'goFlux' function, the input dataframe allows for chamID to be used instead of UniqueID, but when I go to plot the results using flux.plot, it throws an error:
>
> ```
> library(goFlux)
> CO2_flux <- goFlux(LI8100_imp, "CO2dry_ppm")
> CO2_flux_res <- best.flux(CO2_flux)
> CO2_flux_plots <- flux.plot(CO2_flux_res, LI8100_imp, "CO2dry_ppm", shoulder=20,
>                             plot.legend, plot.display, quality.check)
> ```
>
> Error in flux.plot(CO2_flux_res, LI8100_imp, "CO2dry_ppm", shoulder = 20, : 'UniqueID' is required and was not found in 'dataframe'
>
> The argument description for the flux.plot function says the same dataframe can be used for both goFlux and flux.plot. Would it be possible to allow chamID in flux.plot as well?

Hi @shubhamjain15, this is a very good point. I added the option to the flux.plot function to use chamID instead of UniqueID.
