### 2024-03-05 #6393 @shubhamjain15
Thanks @Qepanna, it works for me. Another quick question - While loading the "DX4015.TXT" file, it only loads the first 14 rows upto time - 08:20:35. However, in the txt file, I do see more rows that go upto 08:30:44. Is there any specific reason why those rows were not read or excluded by the import function?

  ```{r}
library(goFlux)
file.path <- system.file("extdata", "DX4015/DX4015.TXT", package = "goFlux")
DX4015_imp <- DX4015_import(inputfile = file.path)
print(nrow(DX4015_imp))
```
[1] 14
