### 2024-03-06 #6393 @Schuch666
> > Hi @Qepanna, You mention in the installation section that "must first install the package devtools", but the user can use another package that provide the same functionality such as the `remotes` package:
> > `remotes::install_github('qepanna/goflux')`
> > can you include a note that include remotes?
>
> Hi @Schuch666, originally it was included that one could use the remotes packages, however I encounterd some errors using this package and I thought it would be best to avoid using it. Would you still recommend that I include the package remotes?

Hello @Qepanna,

I did not have any issues installing using remotes (for me I use remotes for R installations that don't need all the developing tools from devtools because is ligther), but it is up to you to include or not instructions to use remotes. But instead of "must first install devtools" in the installation instructions should be something less restrictive.
