+++ README.md
@@ -145,18 +145,18 @@ After import, the user can choose from two methods for identification of
 measurements:
 
 - **Manual identification of measurements** - based on `start.time`,
-  provided separately in an auxiliary file. The function `obs.win()`
+  provided separately in an auxiliary file. The function `obs.win`
   splits the imported data into a list of data frame (divided by
   `UniqueID`) and creates an observation window around the `start.time`
   to allow for a manual selection of the start and end points of each
-  measurements, using the function `click.peak.loop()`.
+  measurements, using the function `click.peak.loop`.
 - **Automatic selection of measurements** - based on automatic
   recordings of chamber opening and closing from an instrument such as
   the LI-COR Smart Chamber or the GAIATECH Automated ECOFlux chamber.
 
 ### Flux calculation
 
-The function `goFlux()` calculates fluxes from a variety of greenhouse
+The function `goFlux` calculates fluxes from a variety of greenhouse
 gases (CO<sub>2</sub>, CH<sub>4</sub>, N<sub>2</sub>O, NH<sub>3</sub>,
 CO, and H<sub>2</sub>O) using both linear (LM) and non-linear (HM;
 [Hutchinson and Mosier,
@@ -182,9 +182,9 @@ $$\mathbf{Eqn~2}~~~~~~k.max = \frac{LM.flux}{MDF~\times~t}$$
 Where $LM.flux$ and $MDF$ have the same units (nmol or
 µmol·m<sup>-2</sup>·s<sup>-1</sup>) and $t$ is in seconds. Therefore,
 the units of kappa-max is s<sup>-1</sup>. This limit of kappa-max is
-included in the `goFlux()` function, so that the non-linear flux
-estimate cannot exceed this maximum curvature. See below for more
-details about the minimal detectable flux (MDF).
+included in the `goFlux` function, so that the non-linear flux estimate
+cannot exceed this maximum curvature. See below for more details about
+the minimal detectable flux (MDF).
 
 All flux estimates, including the MDF, are multiplied by a $flux.term$
 which is used to correct for water vapor inside the chamber, as well as
@@ -202,8 +202,7 @@ parameters are measured inside the chamber at $t = 0$.
 ### Automatic selection of the best flux estimate
 
 Following flux calculation, the user can select the best flux estimate
-(LM or HM) based on objective criteria, using the `best.flux()`
-function:
+(LM or HM) based on objective criteria, using the `best.flux` function:
 
 - **Assumed non-linearity**: If all criteria are met, the best flux
   estimate is assumed to be the non-linear estimate from the Hutchinson
@@ -218,7 +217,7 @@ function:
   indices of model fit, measurements can be flagged “noisy” using these
   criteria.
 
-In addition, the `best.flux()` function can flag measurements that are
+In addition, the `best.flux` function can flag measurements that are
 below detection limit (MDF and *p-value*), out of bounds (intercept), or
 too short (number of observations).
 
@@ -243,7 +242,7 @@ al., 2018](https://doi.org/10.1371/journal.pone.0200876)).
 
 $$\mathbf{Eqn~4}~~~~~~g-factor = \frac{HM.flux}{LM.flux}$$
 
-With the `best.flux()` function, one can choose a limit at which the HM
+With the `best.flux` function, one can choose a limit at which the HM
 model is deemed to overestimate (*f*<sub>0</sub>). Recommended
 thresholds for the g-factor are \<4 for a flexible threshold, \<2 for a
 medium threshold, or \<1.25 for a more conservative threshold. The
@@ -274,10 +273,10 @@ No best flux estimate is chosen based on MDF.
 The parameter kappa determines the curvature of the non-linear
 regression in the Hutchinson and Mosier model, as shown in equation 1.
 The limit of kappa-max, as calculated in equation 2, is included in the
[…]
