### 2024-03-08 #6393 @Schuch666
Dear all, there is my general comments:

In general the documentation is very good, and the documentation (specially the website) has god additional technical content.

I found some small errors running some of the "import raw data into R" examples but the last version of the package was this errors already fixed.

And a few questions/suggestions:
1. In the paper/documentation there is "identify measurements", can you include more information (identification of the instrument? identification of start/end of measures? variables that are measured? other flags?)
2. For the visualization part (https://qepanna.quarto.pub/goflux/flux2pdf.html), there is a part that change the timezone of R environment, I recommend to include as a tip instead of the example.
3. Can a option to save the window or disable the auto-close of the `click.peak` function? This plot can be useful
4. In the goFlux webpage (https://qepanna.quarto.pub/goflux/goFlux.html), there is a `data(LGR_manID)` that need to be included in the example.
5. I have a recommendation to the `goFlux` functiuon: if the function is called without an argument `gastype`, can be included a menu to choose the variable or a print message with the variable names (or maybe a print of the variables that can be used and an error if this is part of the desing).
6. What is the difference of goFlux and best.flux? Is the first one always combines LM and HM and the second one use some criteria and calculate the flux using LM or HM depending of the results? There is any diagnostic output to inform the user more about the best method or how the output was calculated (some variables on the output, a message or some kind of log file)?
7. In the website (https://qepanna.quarto.pub/goflux/bestflux.html), there is a `data(LGR_manID)` that need to be included in the example.
8. In the website (https://qepanna.quarto.pub/goflux/flux2pdf.html), part of the example is on different box (tip) and the how to calculate `CO2_flux_res` and `CH4_flux_res` is missing, this can be found in previous page but should be better include all steps

I will include some of this suggestions on the issue page
