### 2024-03-08 Qepanna/goFlux#1 @Schuch666
sugestion: goFlux without gastype
In the `goFlux` function, if the function is called without an argument `gastype` (but with a `dataframe`), can be included a menu to choose the variable from the data.frame or a print message with the variable names that can be used?
