### 2024-03-08 Qepanna/goFlux#2 @Schuch666
Suggestion: keep the plots from click.peak
Can an option to save the plot from the pop-up window or disable the auto-close of the `click.peak` function? I think this plot can be interesting to keep
