### 2024-03-08 Qepanna/goFlux#3 @Schuch666
Diagnostic output from goFlux function
There is any diagnostic output to inform the user more about the best method was chosen (ML or HM) or how the output was calculated? some output message, diagnostic variable or a log of the process can provide a better understanding of the results
