### 2024-03-09 #6393 @Schuch666
There is an inconsistency on the naming convention on the functions: You have function with the name `best.flux` (with a dot) `flux2pdf` (a number), `goFlux` (a capital letter), `EGM5_import` (using underline) and `goFlux_nokmax` (using an underline and a capital letter). I suggest that you rename these files to match with only one naming standard.

For the functions for specific instruments: some cases the names are not very informative to the user for LI7820_imp (this is a dataset from LI7820 instrument), and for other cases change the order LI7820_import to import_LI7820 make more sense in the organization of the documentation and help to write the code (but this is up to you to change or not).
