### 2024-03-13 25d0f2d
filter out NAs and negative values, but keep 0
