+++ R/LI7810_import.R
@@ -135,9 +135,9 @@ LI7810_import <- function(inputfile, date.format = "ymd", timezone = "UTC",
       # Standardize column names
       rename(CO2dry_ppm = CO2, CH4dry_ppb = CH4, H2O_ppm = H2O) %>%
       # Remove NAs and negative gas measurements, if any
-      filter(CO2dry_ppm > 0) %>%
-      filter(CH4dry_ppb > 0) %>%
-      filter(H2O_ppm > 0)
+      filter(CO2dry_ppm >= 0) %>%
+      filter(CH4dry_ppb >= 0) %>%
+      filter(H2O_ppm >= 0)
 
     # Keep only useful columns for gas flux calculation
     if(keep_all == FALSE){

[…]
