### 2024-03-15 #6393 @Qepanna
> > > Hi @Qepanna, You mention in the installation section that "must first install the package devtools", but the user can use another package that provide the same functionality such as the `remotes` package:
> > > `remotes::install_github('qepanna/goflux')`
> > > can you include a note that include remotes?
> >
> >
> > Hi @Schuch666, originally it was included that one could use the remotes packages, however I encounterd some errors using this package and I thought it would be best to avoid using it. Would you still recommend that I include the package remotes?
>
> Hello @Qepanna,
>
> I did not have any issues installing using remotes (for me I use remotes for R installations that don't need all the developing tools from devtools because remotes is lighter), but it is up to you to include or not instructions to use remotes (If users still need to avoid remotes to install your package). But instead of "must first install devtools" in the installation instructions should be something less restrictive.
>
> PS: is possible that the issue with your package / remotes is already fixed because both packages may have several updates

I included that both packages can be used and I just found out that the issue can arise with either packages. The issue concerns the API rate limit. There is a troubleshoot section about that in the README file and the webpage. However, it anoys me that users are required to create a GitHub account to avoid that problem. Do you know if there is any way to avoid this issue altogether?
