### 2024-03-15 7c8668d
change error message when goFlux(gastype = NULL)
