### 2024-03-15 #6393 @Qepanna
> Dear all, there is my general comments:
>
> In general the documentation is very good, and the documentation (specially the website) has god additional technical content.
>
> I found some small errors running some of the "import raw data into R" examples but the last version of the package was this errors already fixed.
>
> And a few questions/suggestions:
>
> 1. In the paper/documentation there is "identify measurements", can you include more information (identification of the instrument? identification of start/end of measures? variables that are measured? other flags?)
> 2. For the visualization part (https://qepanna.quarto.pub/goflux/flux2pdf.html), there is a part that change the timezone of R environment, I recommend to include as a tip instead of the example.
> 3. Can an option to save the window or disable the auto-close of the `click.peak` function? This plot can be useful ([Suggestion: keep the plots from click.peak Qepanna/goFlux#2 (comment)](https://github.com/Qepanna/goFlux/issues/2#issue-2176867056))
> 4. In the goFlux webpage (https://qepanna.quarto.pub/goflux/goFlux.html), there is a `data(LGR_manID)` that need to be included in the example.
> 5. I have a recommendation to the `goFlux` functiuon: if the function is called without an argument `gastype`, can be included a menu to choose the variable or a print message with the variable names (or maybe a print of the variables that can be used and an error if this is part of the desing). ([sugestion: goFlux without gastype Qepanna/goFlux#1 (comment)](https://github.com/Qepanna/goFlux/issues/1#issue-2176863327))
> 6. What is the difference of goFlux and best.flux? Is the first one always combines LM and HM and the second one use some criteria and calculate the flux using LM or HM depending of the results? There is any diagnostic output to inform the user more about the best method or how the output was calculated (some variables on the output, a message or some kind of log file)? ([Diagnostic output from goFlux function Qepanna/goFlux#3 (comment)](https://github.com/Qepanna/goFlux/issues/3#issue-2176870036))
> 7. In the website (https://qepanna.quarto.pub/goflux/bestflux.html), there is a `data(LGR_manID)` that need to be included in the example.
> 8. In the website (https://qepanna.quarto.pub/goflux/flux2pdf.html), part of the example is on different box (tip) and the how to calculate `CO2_flux_res` and `CH4_flux_res` is missing, this can be found in previous page but should be better include all steps
> 9. On the website (https://qepanna.quarto.pub/goflux/import.html#mgga-gla131-series) the file used on the examples from UGGA and MGGA are the same, are they suppose to be the same (with identical format)?
>
> I have included 3 issues with my suggestions on the issue page

Dear @Schuch666, thank you for all your suggestions. I have modified the related functions and the webpage accordingly for the most part. Here are my detailed answers to your comments. I also copied those answers on the corresponding issues you oppened.

1. I do not understand the comment. This step is already clearly identified as coming after the import, therefore I cannot add anything about “identification of the instrument”. On the webpage, there is an extensive demonstration about the “identification of start/end of measures”. For the other two suggestions, I don’t understand what you mean. Could you give specific examples of modifications in the text where it would improve the clarity?
2. If you meant instead the demonstration on the page https://qepanna.quarto.pub/goflux/manualID.html, I moved that part to a Tip Callout.
3. I see two options for saving these plots: Either create a new function that will make plots from the output of click.peak, OR if save.plot = TRUE in click.peak, the output of click.peak is now a list of two objects: one data frame and one list of plots. The first option requires a bit more effort, but, from a “user-friendly” point of view, it can help users avoid making mistakes.
On the other hand, these plots are not that useful to keep, since the function flux.plot will essentially return the same plot, but with additional content from the output of goFlux and best.flux.
What do you think is best?
4. I added data(LGR_manID) to the example.
5. I changed the error message for goFlux if gastype = NULL. Now, if no argument is provided to gastype, an error message is returned:
```
Error in goFlux(LGR_manID, gastype = NULL) :
  'gastype' must be one of the following: 'CO2dry_ppm', 'CH4dry_ppb', 'COdry_ppb', 'N2Odry_ppb', 'NH3dry_ppb' or 'H2O_ppm'
```
6. The function goFlux calculates both LM and HM. The function best.flux adds diagnostic columns and the best.flux column, which is the selection between LM and HM. In the description of the best.flux function (?best.flux), under #Value, it says:
“A data.frame identical to the input flux.result (output from goFlux) with the additional columns HM.diagnose, LM.diagnose, best.flux, model and quality.check. For each criteria selected, an additional column is also added to specify the limits used for those criteria (e.g. RMSE.lim, p.val.lim, etc.)”
In the description of the function, it is well detailed how each criterion is evaluated and how the diagnostic is printed in these columns. For example, for the criteria g.limit, it says in the description:
“The g.limit indicates a threshold for the g.factor, which is the ratio between a non-linear flux estimate and a linear flux estimate. With the best.flux function, one can choose a limit at which the HM model is deemed to overestimate f0. Recommended thresholds for the g-factor are <4 for a flexible threshold, <2 for a medium threshold, or <1.25 for a more conservative threshold. The default threshold is g.limit = 2. If the g-factor is above the specified threshold, the best flux estimate will switch to LM instead of HM and give a warning in the columns quality.check and HM.diagnose.”
All of these details are also mentionned on the webpage.
I believe the output from best.flux and the description of the function as they already are fit with the suggestion you gave and no action is required.
7. I added data(LGR_manID) to the example.
8. The tips boxes were rearranged after the example so that it is easier to follow and copy-paste errors should be easily avoided. In addition, all steps of the calculations for each gas were included.
9. Yes, it is normal that they are the same. These two instruments return the same file format. That is also why the function is the same for these two instruments. However, I thought it would be best to make two examples because their precisions are different (argument prec in the function).
