### 2024-03-15 Qepanna/goFlux#2 @Qepanna
> Can an option to save the plot from the pop-up window or disable the auto-close of the `click.peak` function? I think this plot can be interesting to keep

I see two options for saving these plots: Either create a new function that will make plots from the output of click.peak, OR if save.plot = TRUE in click.peak, the output of click.peak is now a list of two objects: one data frame and one list of plots. The first option requires a bit more effort, but, from a “user-friendly” point of view, it can help users avoid making mistakes.

On the other hand, these plots are not that useful to keep, since the function flux.plot will essentially return the same plot, but with additional content from the output of goFlux and best.flux.

What do you think is best?
