### 2024-03-15 Qepanna/goFlux#1 @Qepanna
> In the `goFlux` function, if the function is called without an argument `gastype` (but with a `dataframe`), can be included a menu to choose the variable from the data.frame or a print message with the variable names that can be used?

@Schuch666 I changed the error message for goFlux if gastype = NULL. Now, if no argument is provided to gastype, an error message is returned:

```
Error in goFlux(LGR_manID, gastype = NULL) :
  'gastype' must be one of the following: 'CO2dry_ppm', 'CH4dry_ppb', 'COdry_ppb', 'N2Odry_ppb', 'NH3dry_ppb' or 'H2O_ppm'
```
