### 2024-03-15 Qepanna/goFlux#3 @Qepanna
> There is any diagnostic output to inform the user more about the best method was chosen (ML or HM) or how the output was calculated? some output message, diagnostic variable or a log of the process can provide a better understanding of the results

@Schuch666 The function goFlux calculates both LM and HM. The function best.flux adds diagnostic columns and the best.flux column, which is the selection between LM and HM. In the description of the best.flux function (?best.flux), under #Value, it says:

“A data.frame identical to the input flux.result (output from goFlux) with the additional columns HM.diagnose, LM.diagnose, best.flux, model and quality.check. For each criteria selected, an additional column is also added to specify the limits used for those criteria (e.g. RMSE.lim, p.val.lim, etc.)”

In the description of the function, it is well detailed how each criterion is evaluated and how the diagnostic is printed in these columns. For example, for the criteria g.limit, it says in the description:

“The g.limit indicates a threshold for the g.factor, which is the ratio between a non-linear flux estimate and a linear flux estimate. With the best.flux function, one can choose a limit at which the HM model is deemed to overestimate f0. Recommended thresholds for the g-factor are <4 for a flexible threshold, <2 for a medium threshold, or <1.25 for a more conservative threshold. The default threshold is g.limit = 2. If the g-factor is above the specified threshold, the best flux estimate will switch to LM instead of HM and give a warning in the columns quality.check and HM.diagnose.”

I believe the output from best.flux and the description of the function as they already are fit with the suggestion you gave and no action is required.
