### 2024-03-15 Qepanna/goFlux#1 @Schuch666
This error is more informative

An suggestion for the future is to inform the available variables (if the right columns are present on the input data.frame and if contains any data)
