### 2024-03-15 Qepanna/goFlux#2 @Schuch666
I think a function duplicating the same code is not the best option, the second options looks better (and do not require to change the output format).

And how about the user give only a file.name in case he/she want to save the plot?
