### 2024-03-17 #6393 @Schuch666
Dear all,

@Qepanna
1. This comment is specific to the paper summary "2. identify measurements", replace for "define the start and end points of each measurement and assign a UniqueID", identify measurements is not self explanatory.

For 2-9 everything is fine.

About the issue with remotes and API rate limit, I'm not sure, can you create a issue with the steps to reproduce the error and tag me?

I don't have a reply about the inconsistency / non-explanatory names used on the package and I'm including this as a suggestion on the issue session of the package repository.

Thank you
