### 2024-03-18 #6393 @Qepanna
> I have successfully installed the package and tested out various functions in the documentation. The documentation is very thorough and detailed. It may be helpful to have a separate bare bones "Example" section where it runs through each step (e.g., import, goFlux, best.flux, plotting) for just a single data set. The documentation currently has a fair bit of methodology/equations that could potentially be an Appendix.

Dear @hahsan1, I agree that an overall example is necessary. I had planned to make downloadable scripts for each pathway. Could you open an issue with your suggestion?
