+++ paper/paper.md
@@ -28,9 +28,9 @@ affiliations:
 
 # Summary
 
-This new R package `goFlux` has been developed for calculating greenhouse gas (GHG) flux estimates from static chamber measurements. Compared to previous software developed for the same purpose, the `goFlux` package is not limited to the linear regression approach (LM), but also estimates GHG fluxes from a non-linear regression approach, the Hutchinson and Mosier model (HM). An automatic selection procedure has been implemented in the package to help users select the best flux estimate (LM or HM) based on objective criteria. In addition, this package can be used to import raw data directly downloaded from a vast selection of instruments (LI-COR, LGR, GAIA2TECH, Gasmet, Picarro and PP-Systems).
+This new R package `goFlux` has been developed for calculating greenhouse gas (GHG) flux estimates from static chamber measurements. Compared to previous software developed for the same purpose, the `goFlux` package is not limited to the linear regression approach (LM), but also estimates GHG fluxes from a non-linear regression approach, the Hutchinson and Mosier model (HM). An automatic selection procedure has been implemented in the package to help users select the best flux estimate (LM or HM) based on objective criteria. In addition, this package can be used to import raw data directly downloaded from a vast selection of instruments (LI-COR, LGR, GAIA2TECH, Gasmet, Picarro, Aeris and PP-Systems).
 
-The package is divided into five steps: 1. import raw data into R; 2. identify measurements; 3. calculate GHG flux estimates (LM and HM); 4. automatically select the best flux estimate (LM or HM) based on our default choices of objective criteria; 5. and visually inspect the results on plots that can be saved as pdf. For a detailed protocol on how to use this package, visit the webpage <https://qepanna.quarto.pub/goflux/>.
+The package is divided into five steps: 1. import raw data into R; 2. define the start and end points of each measurement and assign a UniqueID; 3. calculate GHG flux estimates (LM and HM); 4. automatically select the best flux estimate (LM or HM) based on our default choices of objective criteria; 5. and visually inspect the results on plots that can be saved as pdf. For a detailed protocol on how to use this package, visit the webpage [https://qepanna.quarto.pub/](https://qepanna.quarto.pub/goflux/){.uri}
 
 # Statement of need
 
