### 2024-03-18 #6393 @Qepanna
> 1. define the start and end points of each measurement and assign a UniqueID

Dear @Schuch666, I corrected the paper with your suggestion.
