+++ R/imp.UGGA-data.R
@@ -1,19 +1,19 @@
-#' Example gas measurements data from LGR instruments
+#' Example gas measurements data from the LGR GHG analyzer UGGA
 #'
 #' Raw data from the micro-GGA Los Gatos Research greenhouse gas analyzer that
-#' has been processed by the function \code{\link[goFlux]{LGR_import}}
+#' has been processed by the function \code{\link[goFlux]{import.UGGA}}
 #' of this package.
 #'
 #' @docType data
 #'
-#' @usage data(LGR_imp)
+#' @usage data(imp.UGGA)
 #'
 #' @format An object of class \code{"data.frame"}
 #'
 #' @keywords datasets
 #'
 #' @examples
-#' data(LGR_imp)
-#' head(LGR_imp)
+#' data(imp.UGGA)
+#' head(imp.UGGA)
 #'
-"LGR_imp"
+"imp.UGGA"
