+++ man/goFlux.nokmax.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/goFlux_nokmax.R
-\name{goFlux_nokmax}
-\alias{goFlux_nokmax}
+% Please edit documentation in R/goFlux.nokmax.R
+\name{goFlux.nokmax}
+\alias{goFlux.nokmax}
 \title{goFlux (without kappa-max): a user-friendly GHG fluxes calculation tool}
 \usage{
-goFlux_nokmax(
+goFlux.nokmax(
   dataframe,
   gastype,
   H2O_col = "H2O_ppm",
@@ -171,9 +171,9 @@ number of observations: UniqueID X has 59 observations".
 }
 \examples{
 data(manID.UGGA)
-CO2_flux <- goFlux_nokmax(manID.UGGA, "CO2dry_ppm")
-CH4_flux <- goFlux_nokmax(manID.UGGA, "CH4dry_ppb")
-H2O_flux <- goFlux_nokmax(manID.UGGA, "H2O_ppm")
+CO2_flux <- goFlux.nokmax(manID.UGGA, "CO2dry_ppm")
+CH4_flux <- goFlux.nokmax(manID.UGGA, "CH4dry_ppb")
+H2O_flux <- goFlux.nokmax(manID.UGGA, "H2O_ppm")
 
 }
 \references{
