+++ NAMESPACE
@@ -18,6 +18,7 @@ export(click.peak.loop)
 export(flux.plot)
 export(flux2pdf)
 export(goFlux)
+export(goFlux.nokmax)
 export(goFlux_nokmax)
 export(import.DX4015)
 export(import.EGM5)
