+++ DESCRIPTION
@@ -39,6 +39,7 @@ Collate:
     'g.factor.R'
     'k.max.R'
     'goFlux.R'
+    'goFlux.nokmax.R'
     'goFlux_nokmax.R'
     'imp.LI6400-data.R'
     'imp.LI7820-data.R'
