### 2024-03-18 036cf60
Deploying to gh-pages from @ Qepanna/goFlux@1a75c26ff13d6ac54f52c4a80fdf27e4f1832689 🚀
