### 2024-03-18 d654f79
rm goFlux_nokmax
