+++ R/goFlux_nokmax.R
@@ -1,552 +1,0 @@
-#' goFlux (without kappa-max): a user-friendly GHG fluxes calculation tool
-#'
-#' A wrapper function to calculate GHG fluxes from static chamber measurements.
-#' Calculates linear (\code{\link[goFlux]{LM.flux}}) and non-linear
-#' fluxes (Hutchinson and Mosier model; \code{\link[goFlux]{HM.flux}}),
-#' from a variety of greenhouse gasses (
-#' \ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}},
-#' \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}},
-#' \ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}},
-#' \ifelse{html}{\out{NH<sub>3</sub>}}{\eqn{NH[3]}{ASCII}}, CO, and
-#' \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}).
-#'
-#' @param dataframe a data.frame containing gas measurements (see \code{gastype}
-#'                  below), water vapor measurements (see \code{H2O_col} below)
-#'                  and the following columns: \code{UniqueID}, \code{Etime},
-#'                  \code{Vtot}, \code{Area}, \code{Pcham}, \code{Tcham} and
-#'                  \code{flag} (see the parameters \code{Vtot}, \code{Area},
-#'                  \code{Pcham} and \code{Tcham} below for more details).
-#'                  \code{chamID} may be used instead of \code{UniqueID}.
-#' @param gastype character string; specifies which column should be used for the
-#'                flux calculations. Must be one of the following: "CO2dry_ppm",
-#'                "CH4dry_ppb", "COdry_ppb", "N2Odry_ppb", "NH3dry_ppb" or "H2O_ppm".
-#' @param H2O_col character string; specifies which column should be used to
-#'                subtract the effect of water vapor in the chamber space.
-#'                Default is \code{H2O_col = "H2O_ppm"}.
-#' @param prec numerical value; precision of the instruments. Units must be the
-#'             same as \code{gastype}. With the default \code{prec = NULL},
-#'             instrument precision for each gas must be provided in
-#'             \code{dataframe}.
-#' @param Area numerical value; area of the soil surface inside the chamber
-#'             \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}}.
-#'             Alternatively, provide the column \code{Area} in \code{dataframe}
-#'             if \code{Area} is different between samples.
-#' @param Vtot numerical value; total volume inside the chamber, tubes, instruments,
-#'             etc. (L). Alternatively, provide the column \code{Vtot} in
-#'             \code{dataframe} if \code{Vtot} is different between samples. If
-#'             \code{Vtot} is missing, the function will calculate it from
-#'             \code{Area}, \code{Vcham} and \code{offset}.
-#' @param Vcham (optional) numerical value; volume inside the chamber, tubes and
-#'              instruments (L). Alternatively, provide the column \code{Vcham}
-#'              in \code{dataframe} if \code{Vcham} is different between samples.
-#'              \code{Vhcam} is only used if \code{Vtot} is missing.
-#' @param offset (optional) numerical value; height between the soil surface and
[…]
