+++ man/goFlux_nokmax.Rd
@@ -1,213 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/goFlux_nokmax.R
-\name{goFlux_nokmax}
-\alias{goFlux_nokmax}
-\title{goFlux (without kappa-max): a user-friendly GHG fluxes calculation tool}
-\usage{
-goFlux_nokmax(
-  dataframe,
-  gastype,
-  H2O_col = "H2O_ppm",
-  prec = NULL,
-  Area = NULL,
-  offset = NULL,
-  Vtot = NULL,
-  Vcham = NULL,
-  Pcham = NULL,
-  Tcham = NULL,
-  warn.length = 60
-)
-}
-\arguments{
-\item{dataframe}{a data.frame containing gas measurements (see \code{gastype}
-below), water vapor measurements (see \code{H2O_col} below)
-and the following columns: \code{UniqueID}, \code{Etime},
-\code{Vtot}, \code{Area}, \code{Pcham}, \code{Tcham} and
-\code{flag} (see the parameters \code{Vtot}, \code{Area},
-\code{Pcham} and \code{Tcham} below for more details).
-\code{chamID} may be used instead of \code{UniqueID}.}
-
-\item{gastype}{character string; specifies which column should be used for the
-flux calculations. Must be one of the following: "CO2dry_ppm",
-"CH4dry_ppb", "COdry_ppb", "N2Odry_ppb", "NH3dry_ppb" or "H2O_ppm".}
-
-\item{H2O_col}{character string; specifies which column should be used to
-subtract the effect of water vapor in the chamber space.
-Default is \code{H2O_col = "H2O_ppm"}.}
-
-\item{prec}{numerical value; precision of the instruments. Units must be the
-same as \code{gastype}. With the default \code{prec = NULL},
-instrument precision for each gas must be provided in
-\code{dataframe}.}
-
-\item{Area}{numerical value; area of the soil surface inside the chamber
-\ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}}.
-Alternatively, provide the column \code{Area} in \code{dataframe}
-if \code{Area} is different between samples.}
-
-\item{offset}{(optional) numerical value; height between the soil surface and
-the chamber (cm). Alternatively, provide the column \code{offset}
-in \code{dataframe} if \code{offset} is different between samples.
-\code{offset} is only used if \code{Vtot} is missing.}
-
-\item{Vtot}{numerical value; total volume inside the chamber, tubes, instruments,
-etc. (L). Alternatively, provide the column \code{Vtot} in
-\code{dataframe} if \code{Vtot} is different between samples. If
-\code{Vtot} is missing, the function will calculate it from
-\code{Area}, \code{Vcham} and \code{offset}.}
-
-\item{Vcham}{(optional) numerical value; volume inside the chamber, tubes and
-instruments (L). Alternatively, provide the column \code{Vcham}
-in \code{dataframe} if \code{Vcham} is different between samples.
-\code{Vhcam} is only used if \code{Vtot} is missing.}
-
-\item{Pcham}{numerical value; pressure inside the chamber (kPa).
-Alternatively, provide the column \code{Pcham} in \code{dataframe}
-if \code{Pcham} is different between samples. If \code{Pcham} is
-not provided, normal atmospheric pressure (101.325 kPa) is used.}
-
-\item{Tcham}{numerical value; temperature inside the chamber (Celsius).
-Alternatively, provide the column \code{Tcham} in \code{dataframe}
[…]
