+++ NAMESPACE
@@ -19,7 +19,6 @@ export(flux.plot)
 export(flux2pdf)
 export(goFlux)
 export(goFlux.nokmax)
-export(goFlux_nokmax)
 export(import.DX4015)
 export(import.EGM5)
 export(import.G2508)
