### 2024-03-18 #6393 @Qepanna
> Looks like we're close to wrapping up here. Thanks for the timely reviews folks ⚡
>
> @Schuch666 – I can see that there are a few items left unchecked in your review checklist – are you still awaiting changes from the author on these matters or do you believe your feedback has been incorporated and these can now be checked off?

Dear, @arfon, I am only missing two comments to answer:
- Keep plots from click.peak (suggestion from @Schuch666 opened as an issue)
- Create a separate "bare bone" example on the webpage (suggestion from @hahsan1)
