### 2024-03-18 #6393 @hahsan1
I have posted an issue highlighting the suggestion to add a standalone example in the documentation. I have no further comments to add. Great work on the package!
