### 2024-03-20 e1fa101
add the argument save.plots; click.peak2 replaces click.peak and click.peak.loop (deprecated)
