+++ R/click.peak2.R
@@ -1,0 +1,416 @@
+#' Manual identification of start and end of gas measurements
+#'
+#' Identify the start and the end of a measurement by clicking on them in a
+#' scatter plot. To use in a loop with multiple measurements, first apply the
+#' function \code{\link[goFlux]{obs.win}} to identify the observation
+#' window of each measurement and then use the function
+#' \code{\link[goFlux]{click.peak2}} (see examples below).
+#'
+#' @param ow.list list of data.frame; output from the function
+#'                \code{\link[goFlux]{obs.win}}. Must contain the columns
+#'                \code{gastype} (see below), \code{POSIX.time} and \code{UniqueID}.
+#' @param gastype character string; specifies which gas should be displayed on
+#'                the plot to manually select start time and end time of
+#'                measurements. Must be one of the following: "CO2dry_ppm",
+#'                "COdry_ppb", "CH4dry_ppb", "N2Odry_ppb", "NH3dry_ppb" or
+#'                "H2O_ppm". Default is "CO2dry_ppm".
+#' @param sleep numerical value; delay before closing the resulting plot. Grants
+#'              a delay between measurements to visually inspect the output
+#'              before processing the next measurement. Sleep must be shorter
+#'              than 10 seconds. If \code{sleep = NULL}, the  plots will not close.
+#' @param plot.lim numerical vector of length 2; sets the Y axis limits in the
+#'                 plots. Default values are set for a typical gas measurement
+#'                 of "CO2dry_ppm" from soils: \code{plot.lim = c(380,1000)}.
+#' @param seq a numerical sequence that indicates objects in a list. By default,
+#'            \code{seq = NULL} and the function loops through all data frames
+#'            in \code{ow.list}.
+#' @param warn.length numerical value; limit under which a measurement is flagged
+#'                    for being too short (\code{nb.obs < warn.length}). Default
+#'                    value is \code{warn.length = 60}.
+#' @param save.plots character string; a file name without the extension .pdf to
+#'                   save the plots produced with click.peak2. By default,
+#'                   \code{save.plot = NULL} and plots are not saved.
+#'
+#' @returns A list of data.frame, identical to an unlisted version of the input
+#'          \code{ow.list}, with the additional columns \code{flag}, \code{Etime},
+#'          \code{start.time_corr}, \code{end.time_corr} and \code{obs.length_corr}.
+#'
+#' @include goFlux-package.R
+#'
+#' @seealso See also \code{\link[goFlux]{obs.win}} to prepare a list of data.frame.
+#'
+#' @details
+#' The argument \code{plot.lim} is used to remove any data points below and
+#' above the plot limits for a better view of the scatter plot. If the gas
+#' measurements are larger than the minimum or smaller than the maximum plot
+#' limit values, then the plot will automatically zoom in and adjust to those
+#' values. The default plot limits are set for a typical gas measurement of
+#' "CO2dry_ppm" from a soil respiration measurement: \code{plot.lim = c(380,1000)},
+#' where 380 ppm is the minimum plotted concentration, which should be close to
[…]
