+++ R/goFlux-package.R
@@ -49,10 +49,14 @@
 #' @importFrom ggstar geom_star
 #' @importFrom graphics axis.POSIXct
 #' @importFrom graphics identify
+#' @importFrom graphics mtext
+#' @importFrom graphics par
+#' @importFrom graphics text
 #' @importFrom grDevices dev.flush
 #' @importFrom grDevices dev.new
 #' @importFrom grDevices dev.off
 #' @importFrom grDevices pdf
+#' @importFrom grDevices recordPlot
 #' @importFrom grid gpar
 #' @importFrom grid textGrob
 #' @importFrom gridExtra marrangeGrob
@@ -88,6 +92,7 @@
 #' @importFrom tibble add_column
 #' @importFrom tidyr drop_na
 #' @importFrom tidyr pivot_longer
+#' @importFrom toOrdinal toOrdinal
 #' @importFrom utils read.delim
 #' @importFrom utils setTxtProgressBar
 #' @importFrom utils timestamp
