### 2024-03-20 Qepanna/goFlux#2 @Qepanna
@Schuch666 While trying my best to add an option to save plots in the functions click.peak and click.peak.loop, I realized that these two functions were quite unelegant. I then decided to create a new function click.peak2 to replace those two functions and I deprecated them.

The new function click.peak2 can now loop through a list of data frame like click.peak.loop, without the use of lapply, and there is an option to save plots.
