+++ man/click.peak.loop.Rd
@@ -52,7 +52,7 @@ A list of data.frame, split by \code{UniqueID}, identical to the
 }
 \description{
 DEPRECATED: This function will be removed in a future version of the package.
-Use the function \code{\link[goFlux]{click.peak2}} instead.
+Use the function \code{\link[goFlux]{click.peak2}} instead. \cr \cr
 Identify the start and the end of a measurement by clicking on them in a
 scatter plot. To use in a loop with multiple measurements, first use the
 function \code{\link[goFlux]{obs.win}} to identify the observation
