### 2024-03-20 3c12f6c
change default k.min
