+++ R/goFlux.R
@@ -59,7 +59,7 @@
 #' @param warn.length numerical value; limit under which a measurement is
 #'                    flagged for being too short (\code{nb.obs < warn.length}).
 #' @param k.min numerical value; a lower boundary value for kappa in the HM model.
-#'              Default is \code{k.min = 10^-8}
+#'              Default is \code{k.min = 10^-9}
 #'
 #' @details
 #' Flux estimate units are
@@ -80,7 +80,7 @@
 #' estimate cannot exceed this maximum curvature. Inversely, one can set a
 #' minimal threshold for kappa: to allow for a log-like curvature, set
 #' \code{k.min} below 0 (ex. -1), otherwise it should be just above 0
-#' (ex. 10^-8). \code{k.min} cannot be 0 as this would result in a singular gradient.
+#' (ex. 10^-9). \code{k.min} cannot be 0 as this would result in a singular gradient.
 #'
 #' All flux estimates, including the \code{\link[goFlux]{MDF}}, are
 #' obtained from the multiplication of the slope and the
@@ -193,7 +193,7 @@
 goFlux <- function(dataframe, gastype, H2O_col = "H2O_ppm", prec = NULL,
                    Area = NULL, offset = NULL, Vtot = NULL, Vcham = NULL,
                    Pcham = NULL, Tcham = NULL, k.mult = 1,
-                   warn.length = 60, k.min = 10^-8){
+                   warn.length = 60, k.min = 10^-9){
 
   # Check arguments ####
   if(!is.numeric(k.mult)) stop("'k.mult' must be of class numeric")
