### 2024-03-20 9ea18d3
Deploying to gh-pages from @ Qepanna/goFlux@3c12f6cbe9d1d9e69e97b9f28dd8f8441b5dbbc4 🚀
