### 2024-03-22 9ca7d5e
Convert cm3 to L when calculating Vtot from Area and offset
