+++ paper/paper.md
@@ -34,11 +34,11 @@ The package is divided into five steps: 1. import raw data into R; 2. define the
 
 # Statement of need
 
-Non-steady state chambers are widely used for measuring soil greenhouse gas (GHG) fluxes, such as CO~2~, CH~4~, N~2~O, NH~3~, CO, and water vapor (H~2~O). While linear regression (LM) is commonly used to estimate GHG fluxes, this method tends to underestimate the pre-deployment flux (*f*~0~). Indeed, non-linearity is expected when the gas concentration increases or decreases inside a closed chamber, due to changes in gas gradients between the soil and the air inside the chamber. In addition, lateral gas flow and leakage contribute to non-linearity. Many alternative to LM have been developed to try and provide a more accurate estimation of *f*~0~, for instance the method of @hutchinson1981 (HM), which was implemented in the [`HMR`](https://cran.r-project.org/package=HMR) package by @pedersen2010. However, non-linear models have a tendency to largely overestimate some flux measurements, due to an exaggerated curvature. Therefore, the user is expected to decide which method is more appropriate for each flux estimate. With the advent of portable greenhouse gas analyzers (e.g. [Los Gatos Research (ABB) laser gas analyzers](https://new.abb.com/products/measurement-products/analytical/laser-gas-analyzers/laser-analyzers/lgr-icos-portable-analyzers)), soil GHG fluxes have become much easier to measure, and more efficient ways to calculate these flux estimates are needed in order to process large amounts of data. A recent approach was developed by @hüppi2018 to restrict the curvature in the HM model for a more reliable flux estimate. In the HM model, the curvature is controlled by the kappa parameter. @hüppi2018 suggest the use of the kappa.max to limit the maximal curvature allowed in the model (see their R package [`gasfluxes`](https://cran.r-project.org/package=gasfluxes), available on the CRAN). This procedure introduces less arbitrary decisions in the flux estimation process.
[…]
