### 2024-04-02 #6393 @Qepanna
> OK, thanks for the great discussion here and the iterative improvements throughout the review!
>
> @Qepanna – looks like we're very close to being done here. I will circle back here next week, but in the meantime, please give your own paper a final read to check for any potential typos etc.
>
> After that, could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
>
> * The title of the archive is the same as the JOSS paper title
> * That the authors of the archive are the same as the JOSS paper authors
> * I can then move forward with accepting the submission.

Hi @arfon, here is the DOI from Zenodo: https://zenodo.org/doi/10.5281/zenodo.10908262
