### 2024-04-04 95f7c89
import2RData accepts all parameters from import.GAIA
