### 2024-04-04 4dbc784
if Operating Status missing in raw file, default to 2
