+++ R/GAIA_import.R
@@ -76,9 +76,9 @@
 #' \code{Op.stat.col} corresponds to the columns Operating Status of each
 #' chamber. The argument \code{CH.col} indicates a character string preceding
 #' the chamber number for each column of the raw data. For example, the column
-#' "COM5A011C06_OperatingStatus" in the raw data file contains the Operating
+#' "COM5A010C06_OperatingStatus" in the raw data file contains the Operating
 #' Status for the chamber 1 if \code{CH.col = "COM5A0"} and
-#' \code{Op.stat.col = "1C06_OperatingStatus"}. If columns are absent from the
+#' \code{Op.stat.col = "0C06_OperatingStatus"}. If columns are absent from the
 #' raw data, the arguments are ignored.
 #'
 #' Note that this function was designed for the following units in the raw file:
@@ -162,33 +162,35 @@ GAIA_import <- function(inputfile, date.format = "ymd", timezone = "UTC",
                         H2O2.col = "XT3C05_H2O"){
 
   # Check arguments
-  if (missing(inputfile)) stop("'inputfile' is required")
-  if (!is.character(inputfile)) stop("'inputfile' must be of class character")
-  if (length(date.format) != 1) stop("'date.format' must be of length 1")
-  if (!any(grepl(date.format, c("ymd", "dmy", "mdy")))) {
+  if(missing(inputfile)) stop("'inputfile' is required")
+  if(!is.character(inputfile)) stop("'inputfile' must be of class character")
+  if(length(date.format) != 1) stop("'date.format' must be of length 1")
+  if(!any(grepl(date.format, c("ymd", "dmy", "mdy")))) {
     stop("'date.format' must be of class character and one of the following: 'ymd', 'dmy' or 'mdy'")}
-  if (!is.character(timezone)) stop("'timezone' must be of class character")
-  if (save != TRUE & save != FALSE) stop("'save' must be TRUE or FALSE")
-  if (active != TRUE & active != FALSE) stop("'active' must be TRUE or FALSE")
-  if (background != TRUE & background != FALSE) stop("'background' must be TRUE or FALSE")
-  if (length(pivot) != 1) stop("'pivot' must be of length 1")
-  if (!any(grepl(pivot, c("long", "wide")))) {
+  if(!is.character(timezone)) stop("'timezone' must be of class character")
+  if(save != TRUE & save != FALSE) stop("'save' must be TRUE or FALSE")
+  if(active != TRUE & active != FALSE) stop("'active' must be TRUE or FALSE")
+  if(background != TRUE & background != FALSE) stop("'background' must be TRUE or FALSE")
+  if(length(pivot) != 1) stop("'pivot' must be of length 1")
+  if(!any(grepl(pivot, c("long", "wide")))) {
     stop("'pivot' must be of class character and one of the following: 'long' or 'wide'")}
-  if (!is.numeric(flag)) stop("'flag' must be of class numeric")
-  if (!is.character(H2O2.col)) stop("'H2O2.col' must be of class character")
-  if (!is.character(N2O.col)) stop("'N2O.col' must be of class character")
-  if (!is.character(H2O1.col)) stop("'H2O1.col' must be of class character")
-  if (!is.character(CH4.col)) stop("'CH4.col' must be of class character")
-  if (!is.character(CO2.col)) stop("'CO2.col' must be of class character")
-  if (!is.character(Op.stat.col)) stop("'Op.stat.col' must be of class character")
-  if (!is.character(PAR.col)) stop("'PAR.col' must be of class character")
-  if (!is.character(Tcham.col)) stop("'Tcham.col' must be of class character")
[…]
