### 2024-04-05 b871d54
fix Op.stat.col absent
